Rename submodule, add changelog entry
